Extract a portion of diff writing code to separate function
